REF: share methods between BlockManager/ArrayManager
