+++ pandas/core/internals/base.py
@@ -3,20 +3,26 @@ Base class for the internal managers. Both BlockManager and ArrayManager
 inherit from this class.
 """
 from typing import (
+    Any,
     List,
     Optional,
     TypeVar,
 )
 
+import numpy as np
+
 from pandas._typing import (
     DtypeObj,
     Shape,
+    final,
 )
 from pandas.errors import AbstractMethodError
+from pandas.util._validators import validate_bool_kwarg
 
 from pandas.core.dtypes.cast import find_common_type
 
 from pandas.core.base import PandasObject
+from pandas.core.construction import extract_array
 from pandas.core.indexes.api import (
     Index,
     ensure_index,
@@ -119,8 +125,84 @@ class DataManager(PandasObject):
     ) -> T:
         raise AbstractMethodError(self)
 
+    def apply_with_block(
+        self: T,
+        f,
+        align_keys: Optional[List[str]] = None,
+        ignore_failures: bool = False,
+        **kwargs,
+    ) -> T:
+        raise AbstractMethodError(self)
+
+    @final
     def isna(self: T, func) -> T:
         return self.apply("apply", func=func)
+
+    @final
+    def where(self: T, other, cond, align: bool, errors: str) -> T:
+        if align:
+            align_keys = ["other", "cond"]
+        else:
+            align_keys = ["cond"]
+            other = extract_array(other, extract_numpy=True)
+
+        return self.apply_with_block(
+            "where",
+            align_keys=align_keys,
+            other=other,
+            cond=cond,
+            errors=errors,
+        )
+
+    @final
+    def putmask(self: T, mask, new, align: bool = True) -> T:
+        if align:
+            align_keys = ["new", "mask"]
+        else:
+            align_keys = ["mask"]
+            new = extract_array(new, extract_numpy=True)
+
+        return self.apply_with_block(
+            "putmask",
+            align_keys=align_keys,
+            mask=mask,
+            new=new,
+        )
+
+    @final
+    def fillna(self: T, value, limit, inplace: bool, downcast) -> T:
+        return self.apply_with_block(
+            "fillna", value=value, limit=limit, inplace=inplace, downcast=downcast
+        )
+
+    @final
+    def downcast(self: T) -> T:
+        return self.apply_with_block("downcast")
+
+    @final
+    def replace(self: T, to_replace, value, inplace: bool, regex: bool) -> T:
+        assert np.ndim(value) == 0, value
+        return self.apply_with_block(
+            "replace", to_replace=to_replace, value=value, inplace=inplace, regex=regex
+        )
+
+    def replace_list(
+        self: T,
+        src_list: list[Any],
+        dest_list: list[Any],
+        inplace: bool = False,
+        regex: bool = False,
+    ) -> T:
+        """ do a list replace """
+        inplace = validate_bool_kwarg(inplace, "inplace")
+
+        return self.apply_with_block(
+            "_replace_list",
+            src_list=src_list,
+            dest_list=dest_list,
+            inplace=inplace,
+            regex=regex,
+        )
 
 
 class SingleDataManager(DataManager):
